Use the JSON version of the RPC API if possible.
If jsonrpclib is available, then use the JSON version of the RPC API
rather than the XML one.

The XML version seems to lose \n (e.g. in descriptions) for some
reason. It’s not clear where this bug comes from - most likely the
xmlrpc plug-in for Trac.

I have not tested this extensively, but general retrieval of ticket
details does work, so everything else ought to be fine in theory!