Handle dates in the response.
The Trac XMLRPC plug-in has a custom format for returning types that
JSON cannot handle (specifically datetime and binary).

Convert any datetime objects that are received to actual
datetime.datetime objects.

This could be expanded to binary objects as well, but we aren’t using
them at the moment, and I’m not sure what the destination type would be
(bytestring?) so leave that.