Add a natural language support
Adds a new command "show" that can interpret
natural language and convert it to a trac query.
The command parses the semantic tree extracted
with the `spacy` trying to identify all the name,
operator, value triplets and build a query.

Cover partial, exact, starting, ending matches and
always check for negations in the sentence.

Because most of the values are already known,
add the ability to add them to the configuration
file, so the interpreter is more accurate.

If everything fails, support quoting to make the
user able to signal complex string as values.