Make a special `status+` attribute.
In pretty much all Trac workflows, resolution is only relevant when the
status is `closed`. This means that when a ticket is not closed, you
are never interested in the resolution, so wouldn’t want to include it
in the fields. However, when the status is closed, the resolution is
very interesting.

Add a special attribute `status+` that is the same as `status` when the
ticket is not closed, but has the resolution appended when it is closed.

Change the default display fields to use `status+` rather than `status`.